refactor(form): extract the form editor into editor.js
